Implemented major part of chess library with tests.
I'm going to start implementing the solution now.
I'm planning to do it in C#, to use a pgn parser for .NET.
I plan to avoid Java even it has library support, in case the find algorithm is performance critical.
Java does not have value types, so C# is chosen.
Most datatypes in the solution are going to be immutable, to avoid mutation bugs and allow for easier paralellisation.
Haven't decided how many levels there should be, and haven't checked the database for how much data there is to download.